requirements.txt
Django~=5.0.5
djangorestframework~=3.15.1
python-dotenv~=1.0.1
asgiref==3.7.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cachetools==5.3.3
django-debug-toolbar==4.4.2
elasticsearch==7.17.9
pillow==10.3.0
django-redis==5.4.0
uwsgi==2.0.26